mj update

+++ internal/event/join_event.go
@@ -9,9 +9,9 @@ import (
 type JoinEvent struct {
 	Ticket string `json:"Ticket"`
 	Client net.Conn
-	//Pending  core.DataBuffer
 	SystemId int64
-	Flag     int32
+	Flag     int64
+
 	EventObj
 }
 
@@ -44,7 +44,7 @@ func (s *JoinEvent) Read(buff core.DataBuffer) error {
 		return err
 	}
 	s.Ticket = ticket
-	flag, err := buff.ReadInt32()
+	flag, err := buff.ReadInt64()
 	if err != nil {
 		return err
 	}
@@ -56,7 +56,7 @@ func (s *JoinEvent) Write(buff core.DataBuffer) error {
 	if err := buff.WriteString(s.Ticket); err != nil {
 		return err
 	}
-	if err := buff.WriteInt32(s.Flag); err != nil {
+	if err := buff.WriteInt64(s.Flag); err != nil {
 		return err
 	}
 	return nil
